+++ pmaw/PushshiftAPI.py
@@ -1,13 +1,4 @@
-import copy
-from concurrent.futures import ThreadPoolExecutor, as_completed
-import json
-from collections import deque
-import logging
-
 from pmaw.PushshiftAPIBase import PushshiftAPIBase
-
-log = logging.getLogger(__name__)
-
 
 class PushshiftAPI(PushshiftAPIBase):
     def __init__(self, *args, **kwargs):
@@ -26,6 +17,7 @@ class PushshiftAPI(PushshiftAPIBase):
             jitter (str, optional) - Jitter to use with backoff, defaults to None, options are None, full, equal, decorr
             checkpoint (int, optional) - Size of interval in batches to print a checkpoint with stats, defaults to 10
             file_checkpoint (int, optional) - Size of interval in batches to cache responses when using mem_safe, defaults to 20
+            praw (praw.Reddit, optional) - Used to enrich the Pushshift items retrieved with metadata directly from Reddit
         """
         super().__init__(*args, **kwargs)
 
@@ -35,25 +27,30 @@ class PushshiftAPI(PushshiftAPIBase):
 
         Input:
             ids (str, list) - Submission id(s) to return the comment ids of
-            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 1000, maximum 1000.
+            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 500, maximum 500.
             mem_safe (boolean, optional) - If True, stores responses in cache during operation, defaults to False
             safe_exit (boolean, optional) - If True, will safely exit if interrupted by storing current responses and requests in the cache. Will also load previous requests / responses if found in cache, defaults to False
+            cache_dir (str, optional) - An absolute or relative folder path to cache responses in when mem_safe or safe_exit is enabled
         Output:
             Response generator object
         """
         kwargs['ids'] = ids
-        return self._search(kind='submission_comment_ids', **kwargs)
+        if('filter_fn' in kwargs):
+            raise ValueError('filter_fn not supported for search_submission_comment_ids')
+        return self._search( kind='submission_comment_ids', **kwargs)
 
     def search_comments(self, **kwargs):
         """
         Method for searching comments, returns an array of comments
 
         Input:
-            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 1000, maximum 1000.
+            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 500, maximum 500.
             max_results_per_request (int, optional) - Maximum number of items to return in a single non-id based request, defaults to 100, maximum 100.
             mem_safe (boolean, optional) - If True, stores responses in cache during operation, defaults to False
             search_window (int, optional) - Size in days for search window for submissions / comments in non-id based search, defaults to 365
             safe_exit (boolean, optional) - If True, will safely exit if interrupted by storing current responses and requests in the cache. Will also load previous requests / responses if found in cache, defaults to False
+            filter_fn (function, optional) - A function used for custom filtering the results before saving them. Accepts a single comment parameter and returns False to filter out the item, otherwise returns True.
+            cache_dir (str, optional) - An absolute or relative folder path to cache responses in when mem_safe or safe_exit is enabled
         Output:
             Response generator object
         """
@@ -64,11 +61,13 @@ class PushshiftAPI(PushshiftAPIBase):
         Method for searching submissions, returns an array of submissions
 
         Input:
-            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 1000, maximum 1000.
+            max_ids_per_request (int, optional) - Maximum number of ids to use in a single request, defaults to 500, maximum 500.
             max_results_per_request (int, optional) - Maximum number of items to return in a single non-id based request, defaults to 100, maximum 100.
             mem_safe (boolean, optional) - If True, stores responses in cache during operation, defaults to False
             search_window (int, optional) - Size in days for search window for submissions / comments in non-id based search, defaults to 365
             safe_exit (boolean, optional) - If True, will safely exit if interrupted by storing current responses and requests in the cache. Will also load previous requests / responses if found in cache, defaults to False
+            filter_fn (function, optional) - A function used for custom filtering the results before saving them. Accepts a single submission parameter and returns False to filter out the item, otherwise returns True.
+            cache_dir (str, optional) - An absolute or relative folder path to cache responses in when mem_safe or safe_exit is enabled
         Output:
             Response generator object
         """