+++ pmaw/Request.py
@@ -3,12 +3,17 @@ import copy
 import datetime as dt
 from collections import deque
 import warnings
+from threading import Event
+import signal
+import time
+
+from praw.exceptions import RedditAPIException
 
 from pmaw.Cache import Cache
 from pmaw.utils.slices import timeslice, mapslice
+from pmaw.utils.filter import apply_filter
 from pmaw.Response import Response
-from threading import Event
-import signal
+
 
 log = logging.getLogger(__name__)
 
@@ -16,9 +21,9 @@ log = logging.getLogger(__name__)
 class Request(object):
     """Request: Handles request information, response saving, and cache usage."""
 
-    def __init__(self, payload, kind, max_results_per_request, max_ids_per_request, mem_safe, safe_exit, cache_dir=None):
+    def __init__(self, payload, filter_fn, kind, max_results_per_request, max_ids_per_request, mem_safe, safe_exit, cache_dir=None, praw=None):
         self.kind = kind
-        self.max_ids_per_request = min(1000, max_ids_per_request)
+        self.max_ids_per_request = min(500, max_ids_per_request)
         self.max_results_per_request = min(100, max_results_per_request)
         self.safe_exit = safe_exit
         self.mem_safe = mem_safe
@@ -26,7 +31,35 @@ class Request(object):
         self.payload = payload
         self.limit = payload.get('limit', None)
         self.exit = Event()
+        self.praw = praw
+        self._filter = filter_fn
 
+        if filter_fn is not None and not callable(filter_fn):
+            raise ValueError('filter_fn must be a callable function')
+
+        if safe_exit and self.payload.get('before', None) is None:
+            # warn the user not to use safe_exit without setting before,
+            # doing otherwise will make it impossible to resume without modifying 
+            # future query to use before value from first run
+            before = int(dt.datetime.now().timestamp())
+            payload['before'] = before
+            warnings.warn(f'Using safe_exit without setting before value is not recommended. Setting before to {before}')
+
+        if self.praw is not None:
+            if safe_exit:
+                raise NotImplementedError('safe_exit is not implemented when PRAW is used for metadata enrichment')
+
+            self.enrich_list = deque()
+            
+            if not kind == 'submission_comment_ids' :
+                # id filter causes an error for submission_comment_ids endpoint
+                self.payload['filter'] = 'id'
+
+            if kind == "submission":
+                self.prefix = "t3_"
+            else:
+                self.prefix = "t1_"
+            
         if 'ids' not in self.payload:
             # add necessary args
             self._add_nec_args(self.payload)
@@ -54,15 +87,62 @@ class Request(object):
         try:
             getattr(signal, 'SIGHUP')
             sigs = ('TERM', 'HUP', 'INT')
-        except Exception:
+        except AttributeError:
             sigs = ('TERM', 'INT')
 
         for sig in sigs:
             signal.signal(getattr(signal, 'SIG'+sig), self._exit)
 
+    def _enrich_data(self):
+        # create batch of fullnames up to 100
+        fullnames = []
+        while len(fullnames) < 100:
+            try:
+                fullnames.append(self.enrich_list.popleft())
+            except IndexError:
+                break
+        
+        # exit loop if nothing to enrich
+        if len(fullnames) == 0:
+            return
+        
+        try:
+            # TODO: may need to change praw usage based on multithread performance
+            resp_gen = self.praw.info(fullnames=fullnames)
+            praw_data = [vars(obj) for obj in resp_gen]
+            results = self._apply_filter(praw_data)
+            self.resp.responses.extend(results)
+            
+        except RedditAPIException:
+            self.enrich_list.extend(fullnames)
+
+    def _idle_task(self, interval):
+        start = time.time()
+        current = time.time()
+
+        if self.praw:
+            # make multiple enrich requests based on sleep interval
+            while current - start < interval and len(self.enrich_list) > 0:
+                
+                self._enrich_data()
+                
+                current = time.time()
+
+        current = time.time()
+        diff = (current - start)
+
+        if diff < interval and diff >= 0:
+            time.sleep(interval-diff)
+
     def save_cache(self):
         # trim extra responses
         self.trim()
+
+        # enrich if needed
+        if self.praw:
+            while len(self.enrich_list) > 0:
+                self._enrich_data()
+
         if self.safe_exit and not self.limit == None and (self.limit == 0 or self.exit.is_set()):
             # save request info to cache
             self._cache.save_info(req_list=self.req_list,
@@ -75,12 +155,30 @@ class Request(object):
     def _exit(self, signo, _frame):
         self.exit.set()
 
+    def _apply_filter(self, results):
+        # apply user defined filter function before storing
+        if(self._filter is not None):
+            return apply_filter(results, self._filter)
+        else:
+            return results    
+
     def save_resp(self, results):
+        # dont filter results before updating limit: limit is the max number of results
+        # extracted from Pushshift, filtering can reduce the results < limit
         if self.kind == 'submission_comment_ids':
             self.limit -= 1
         else:
             self.limit -= len(results)
-        self.resp.responses.extend(results)
+            
+        if self.praw:
+            # save fullnames of objects to be enriched with metadata by PRAW
+            if self.kind == 'submission_comment_ids':
+                self.enrich_list.extend([self.prefix+res for res in results])
+            else:
+                self.enrich_list.extend([self.prefix+res['id'] for res in results])
+        else:
+            results = self._apply_filter(results)
+            self.resp.responses.extend(results)
 
     def _add_nec_args(self, payload):
         """Adds arguments to the payload as necessary."""
@@ -187,7 +285,15 @@ class Request(object):
                 payload['ids'] = list(payload['ids'])
 
     def trim(self):
-        if self.limit and self.limit < 0:
-            log.debug(f'Trimming {self.limit*-1} requests')
-            self.resp.responses = self.resp.responses[:self.limit]
-            self.limit = 0
+        if self.limit:
+            if self.praw:
+                while self.limit < 0:
+                    try:
+                        self.enrich_list.pop()
+                        self.limit += 1
+                    except IndexError as exc:
+                        break
+            if self.limit < 0:
+                log.debug(f'Trimming {self.limit*-1} requests')
+                self.resp.responses = self.resp.responses[:self.limit]
+                self.limit = 0